Sorgenti Python/requirements.txt
pyserial
PyQt6
pyqtgraph
